Improve the initrd design
Now each directory contains a list of offsets of each file
which makes the file mapping easier when the start of the file
from initrd start can be calculated using this offset.

Currently the code doesn't do much with this but in the future
the design will improve and this change will have a more significant
impact. File meta data must be stored separately for efficiency.
Right now large files would exhaust the available memory in an instant.

Fixed also another bug where files at the end of initrd would not load
correctly due to their memory not being mapped. Currently the directory
which contains the file shall map all needed memory.