Fix slab cache allocation bug
The cache_t object is allocated from heap which is in constant use.
vfs_path_lookup() modified some memory and returned it back to
heap which was then allocated by cache_create() but was not zeroed out.

Later on cache_alloc_entry() would use c->free_chunks pointer that
pointed to garbage and caused a page fault