mm: Add boot memory allocator
This is allocator is used only during boot and it's created
because PFA needs heap and slab to initialize themselves.

This implementation is pretty crappy tbh but right now the need for
bootmem allocator is minimal so I don't want to use too much time on
it right now. In the future it'll improve if/when the need arises.