arch/x86_64: Add percpu variables
Each CPU has its own copy of the variables and this is an essential
feature for f.ex. multiprocessor scheduling.

Currently only the get_* macros are useful but kernel preemption is
a planned feature so it is very much encouraged to pair every get_*
with put_*.

doc/SMP.md contains more information about percpu variables