sched: Implement round robin for all cores
From now on the scheduling is done a little differently: Previously
PIT was phased to throw an interrupt very N milliseconds but now when
we're using LAPIC timer that fires every millisecond, we can call the
scheduler from the tick handler.

Every time __tmr_handler() is called, it will call both tick_inc()
and sched_tick(). sched_tick() will update the runtime statistics
of currently running task and if the task has run long enough,
it will be marked and rescheduled.

This commit doesn't introduce support for priorities or SMP load balancing