sched: Redesign context switching
The previous context switching code was suboptimal as each switch
was deemed to return back to user mode with no sleeping in the kernel
between switches. Using that model, creating wait queues would've become
so messy that it was a no-go.

This new context switch works the same way as the old one up until
we're making a decision where to return. The old one picked a task and jumped
right back to user land, but this new firstly switches kernel stacks and
unwinds the kernel stack until the code reaches the place where it
originally reached the kernel [probably __tmr_handler()].

This makes it possible to call sched_switch() for example when waiting
on an input from pipe. The task is put to wait queue and when an
interesting action happens, the task is woken up and it can continue
execution from where it left of.