sched: Increase kernel stack size to 4096 bytes
The code actually allocated one page of stack for each thread but
for some reason I forgot to update the KSTACK_SIZE variable.

Spent two days figuring out why code crashes only when optimizations
are turned off which turned out to be a simple kernel stack overflow
because GCC decided not to inline functions when -O0 is set.