sched: Add proper support for x86_64 scheduling
This is the last commit introducing x86_64-related fixes to the
scheduling subsystem. The whole thing was rewritten because the
pre-x86_64 version of the subsystem was trash and barely even worked.

Now yielding in kernel mode works, waiting for I/O works and
initializing APs work so all and all this is quite ready.

What is still missing is the the actual multicore scheduling and
load balancing but those are not introduced in this branch.