arch/x86_64: Do not map kernel to lower half of PML4
Keep kernel in the higher half and reserve the lower half for user
applications and memory-mapped devices.

This patch actually introduced some ugly changes. micael needs some
kind of framework for memory-mapped device support and right now
(because the kernel is no longer part of lower-half) we need to manually
map Local APIC and I/O APIC each time address space is duplicated which
is ugly.

There also arose a need, because lower-half is no longer mapped, for
memory allocator that returns virtual instead of physical pages.
One way to do this is to create MM_VIRTUAL allocation flag that is
passed to the buddy memory allocator and when it is returning the block,
it checks for this flag and returns higher-half address instead.