lib: Create sliding sampler
Sliding sampler is a type of sampler that only keeps samples from
f.ex. past 10 seconds and returns the average of samples from
that time window.

It is useful for example for storing runtime heuristics of running
programs. It is implemented as a very simpel ring buffer that decays
old values as they're overwritten and from which a time average of
samples can be queried with different time granularities