kernel: Add ACPICA to micael
Add ACPICA so that micael has proper ACPI support. This is mostly
needed for PCI IRQ routing info but it'll be useful for power
management too in the far future

I really don't like ACPI nor am I really hyped about adding
ACPICA to the project either because it's just a huge amount of
code I have no understanding of but I guess it'll have to do for.
Maybe in the very distant future I'll get rid of it and implement
my own AML parser...

Many thanks to Intel for providing ACPICA