mm/heap: Reimplement kernel heap
Split the heap into two arena categories: normal and high priority.

Normal arenas are used for any kind of memory allocation and the
high-priority arena is used to satisfy critical allocations, for
example, an allocation to create a new normal-priority arena.

The implementation is not perfect but it'll have to do for now as I'll
be revising the whole memory manager implementation (again) in couple of
months to make it more robust.