net: Introduce packet_t
This struct encapsulates the whole packet and contains sub-structs
that point to specific parts of the frame such as IPv4 and UDP
headers and payload.

NIC makes a copy of the received Ethernet frame and then this copy
is passed on in the callstack to the final destination that then
frees the memory once the packet has been handled.

Currently only receive stack uses packet_t but in the future, the send
stack will also use packet_t to represent data to remove unnecessary
copies.